refactor: replace ExtensionState with utility function to check for editor and document
