+++ src/dspf-edit.utils/dspf-edit.helper.ts
@@ -664,3 +664,76 @@ export function generateIfDds(
         debounceUpdate(treeProvider, ExtensionState.lastDdsDocument);
     };
 };
+
+/**
+ * Retrieves the current DDS editor and its associated document from extension state.
+ * Validates that both are available before proceeding.
+ * @returns An object containing the active DDS editor and document, or undefined if not found.
+ */
+export function checkForEditorAndDocument() : { editor : vscode.TextEditor | undefined, document :vscode.TextDocument | undefined } {
+
+    const editor = ExtensionState.lastDdsEditor;
+    const document = editor?.document ?? ExtensionState.lastDdsDocument;
+    if (!document || !editor) {
+        vscode.window.showErrorMessage('No DDS editor found.');
+        return { editor: undefined, document: undefined};
+    };
+
+    return { editor, document };
+};
+
+/**
+ * Parses indicators from a DDS line at positions 7-9, 10-12, 13-15.
+ * @param lineText - The DDS line text
+ * @returns Array of indicator codes found
+ */
+export function parseIndicatorsFromLine(lineText: string): string[] {
+    const indicators: string[] = [];
+
+    // Check positions 7-9, 10-12, 13-15 (0-based: 6-8, 9-11, 12-14)
+    const positions = [6, 9, 12];
+
+    for (const pos of positions) {
+        if (lineText.length > pos + 2) {
+            const indicator = lineText.substring(pos, pos + 3).trim();
+            if (indicator && /^N?[0-9]{1,2}$/.test(indicator)) {
+                indicators.push(indicator);
+            };
+        };
+    };
+
+    return indicators;
+};
+
+/**
+ * Groups consecutive line numbers together for more efficient batch deletion.
+ * @param lines - Array of line indices (should be sorted)
+ * @returns Array of arrays, where each sub-array contains consecutive line numbers
+ */
+export function groupConsecutiveLines(lines: number[]): number[][] {
+    if (lines.length === 0) return [];
+    
+    // Ensure lines are sorted
+    const sortedLines = [...lines].sort((a, b) => a - b);
+    const groups: number[][] = [];
+    let currentGroup: number[] = [sortedLines[0]];
+    
+    for (let i = 1; i < sortedLines.length; i++) {
+        const currentLine = sortedLines[i];
+        const previousLine = sortedLines[i - 1];
+        
+        if (currentLine === previousLine + 1) {
+            // Consecutive line - add to current group
+            currentGroup.push(currentLine);
+        } else {
+            // Non-consecutive line - start new group
+            groups.push(currentGroup);
+            currentGroup = [currentLine];
+        };
+    };
+    
+    // Don't forget the last group
+    groups.push(currentGroup);
+    
+    return groups;
+};
